Switch tab view into ticketView component
