+++ src/routes/Group/components/GroupView.js
@@ -1,8 +1,7 @@
 import React from "react";
 import "./GroupView.scss";
-import { Divider, Label, Grid, Header, Button, Confirm, Icon, Segment, Card, Feed, Menu } from 'semantic-ui-react';
+import { Divider, Label, Grid, Header, Button, Confirm, Icon, Segment, Card, Feed } from 'semantic-ui-react';
 import _ from 'lodash';
-import moment from 'moment';
 import InlineEditable from '../../../components/shared/InlineEditable';
 import classNames from 'classnames';
 import TicketView from './TicketView';
@@ -26,8 +25,7 @@ class GroupView extends React.Component {
       "handleAddSubscribers",
       "handleProcessRequest",
       "handleEditPagingInterval",
-      "handleTicketAcknowledgement",
-      "handleTabClick");
+      "handleTicketAcknowledgement");
   }
 
   toggleSelectOnCall(index) {
@@ -99,10 +97,6 @@ class GroupView extends React.Component {
 
   handleTicketAcknowledgement(ticketId) {
     this.props.acknowledgeTicket(ticketId);
-  }
-
-  handleTabClick() {
-    this.setState({ showOpenTickets: !this.state.showOpenTickets });
   }
 
   componentWillMount() {
@@ -178,45 +172,6 @@ class GroupView extends React.Component {
     );
   };
 
-  renderTickets() {
-    return !_.isNil(this.props.openTickets) && this.props.openTickets.map((ticket, index) => {
-      const timeSinceLastUpdate = moment(
-        ticket.updatedAt ?
-        ticket.updatedAt :
-        ticket.createdAt).fromNow();
-
-      const hasBeenPaged = _.some(ticket.actions, (action) =>
-        action.user && action.user._id == this.props.user._id
-      );
-
-      return _.isObject(ticket) && (
-        <Card key={index} color={hasBeenPaged ? 'red' : 'green'} fluid>
-          <Card.Content>
-            <Card.Header>
-              {ticket.metadata.title}
-            </Card.Header>
-            <Feed>
-              <Feed.Event>
-                <Feed.Content>
-                  <Feed.Date content={timeSinceLastUpdate} />
-                  <Feed.Summary>
-                    {ticket.metadata.description}
-                  </Feed.Summary>
-                  <Button
-                    className="action-button"
-                    onClick={() => this.handleTicketAcknowledgement(ticket._id)}
-                  >
-                    Acknowledge Ticket
-                  </Button>
-                </Feed.Content>
-              </Feed.Event>
-            </Feed>
-          </Card.Content>
-        </Card>
-      );
-    });
-  }
-
   renderAdmin() {
     return this.props.group.joinRequests
           && this.props.group.joinRequests.length > 0
@@ -269,15 +224,6 @@ class GroupView extends React.Component {
     );
   }
 
-  renderMenu() {
-    return (
-      <Menu attached="top" tabular>
-        <Menu.Item name='openTickets' active={this.state.showOpenTickets} onClick={this.handleTabClick} />
-        <Menu.Item name='ticketLogs' active={!this.state.showOpenTickets} onClick={this.handleTabClick} />
-      </Menu>
-    );
-  }
-
   render() {
     const leaveButton = this.props.group && _.find(this.props.group.users, (user) => user._id === this.props.user._id) && (
       <Button className="action-button" onClick={this.toggleConfirm}>Leave Group</Button>
@@ -318,14 +264,12 @@ class GroupView extends React.Component {
           </Grid.Column>
         </Grid>
         {this.renderAdmin()}
-        {this.renderMenu()}
-
-        <Segment attached='bottom'>
-          {this.state.showOpenTickets ?
-            this.renderTickets() :
-            <TicketView tickets={this.props.tickets}/>
-          }
-        </Segment>
+        <TicketView
+          user={this.props.user}
+          tickets={this.props.tickets}
+          openTickets={this.props.openTickets}
+          handleTicketAcknowledgement={this.handleTicketAcknowledgement}
+        />
       </div>
     );
   }
